requirements.txt
flask
flask-cors
mysql-connector-python
opencv-python
mediapipe
numpy
face-recognition
ultralytics
